requirements.txt
python-telegram-bot==20.7
pdfplumber==0.11.9
pypdf==5.9.0
reportlab==4.4.10
scikit-learn==1.4.0
sentence-transformers==2.6.0
python-dotenv==1.0.0
aiohttp==3.9.3
numpy==1.26.4
